+++ src/backend/gpu_backend_conv-inl.h
@@ -11,25 +11,25 @@ void Backend<GPUTensor, DType>::Convolution2DForwardFunc(
   // shape decode
   // input
   const Shape& input_shape = input->shape();
-  int batch_size = input_shape[0];
-  int input_channel = input_shape[1];
-  int input_height = input_shape[2];
-  int input_width = input_shape[3];
+  size_t batch_size = input_shape[0];
+  size_t input_channel = input_shape[1];
+  size_t input_height = input_shape[2];
+  size_t input_width = input_shape[3];
   // filter
   const Shape& filter_shape = filter->shape();
-  int filter_height = filter_shape[2];
-  int filter_width = filter_shape[3];
+  size_t filter_height = filter_shape[2];
+  size_t filter_width = filter_shape[3];
   // output
   const Shape& output_shape = output->shape();
-  int output_channel = output_shape[1];
-  int output_height = output_shape[2];
-  int output_width = output_shape[3];
+  size_t output_channel = output_shape[1];
+  size_t output_height = output_shape[2];
+  size_t output_width = output_shape[3];
 
-  int batch_input_offset = 0;
-  int batch_output_offset = 0;
-  int dim_left = output_channel;
-  int dim_right = output_height * output_width;
-  int dim_common = input_channel * filter_height * filter_width;
+  size_t batch_input_offset = 0;
+  size_t batch_output_offset = 0;
+  size_t dim_left = output_channel;
+  size_t dim_right = output_height * output_width;
+  size_t dim_common = input_channel * filter_height * filter_width;
 #ifdef BLITZ_DEVELOP
   LOG(INFO) << "dim left: " << dim_left;
   LOG(INFO) << "dim right: " << dim_right;
@@ -51,7 +51,7 @@ void Backend<GPUTensor, DType>::Convolution2DForwardFunc(
       const_cast<DType*>(input->data()), output->data(),
       const_cast<DType*>(filter->data()), "forward");
   } else {
-    for (int batch_index = 0; batch_index < batch_size; ++batch_index) {
+    for (size_t batch_index = 0; batch_index < batch_size; ++batch_index) {
 #ifdef BLITZ_PERFORMANCE
       cudaEventRecord(start);
 #endif
@@ -119,25 +119,25 @@ void Backend<GPUTensor, DType>::Convolution2DBackwardFunc(
   // shape decode
   // input
   const Shape& input_shape = input->shape();
-  int batch_size = input_shape[0];
-  int input_channel = input_shape[1];
-  int input_height = input_shape[2];
-  int input_width = input_shape[3];
+  size_t batch_size = input_shape[0];
+  size_t input_channel = input_shape[1];
+  size_t input_height = input_shape[2];
+  size_t input_width = input_shape[3];
   // filter
   const Shape& filter_shape = filter->shape();
-  int filter_height = filter_shape[2];
-  int filter_width = filter_shape[3];
+  size_t filter_height = filter_shape[2];
+  size_t filter_width = filter_shape[3];
   // output
   const Shape& output_shape = output->shape();
-  int output_channel = output_shape[1];
-  int output_height = output_shape[2];
-  int output_width = output_shape[3];
+  size_t output_channel = output_shape[1];
+  size_t output_height = output_shape[2];
+  size_t output_width = output_shape[3];
 
-  int batch_input_offset = 0;
-  int batch_output_offset = 0;
-  int dim_left = output_height * output_width;
-  int dim_right = input_channel * filter_height * filter_width;
-  int dim_common = output_channel;
+  size_t batch_input_offset = 0;
+  size_t batch_output_offset = 0;
+  size_t dim_left = output_height * output_width;
+  size_t dim_right = input_channel * filter_height * filter_width;
+  size_t dim_common = output_channel;
 #ifdef BLITZ_DEVELOP
   LOG(INFO) << "dim left: " << dim_left;
   LOG(INFO) << "dim right: " << dim_right;
@@ -160,7 +160,7 @@ void Backend<GPUTensor, DType>::Convolution2DBackwardFunc(
       input->data(), const_cast<DType*>(output->data()),
       const_cast<DType*>(filter->data()), "backward");
   } else {
-    for (int batch_index = 0; batch_index < batch_size; ++batch_index) {
+    for (size_t batch_index = 0; batch_index < batch_size; ++batch_index) {
       #ifdef BLITZ_PERFORMANCE
       cudaEventRecord(start);
       #endif
@@ -227,25 +227,25 @@ void Backend<GPUTensor, DType>::Convolution2DUpdateFunc(
   // shape decode
   // input
   const Shape& input_shape = input->shape();
-  int batch_size = input_shape[0];
-  int input_channel = input_shape[1];
-  int input_height = input_shape[2];
-  int input_width = input_shape[3];
+  size_t batch_size = input_shape[0];
+  size_t input_channel = input_shape[1];
+  size_t input_height = input_shape[2];
+  size_t input_width = input_shape[3];
   // filter
   const Shape& filter_shape = update->shape();
-  int filter_height = filter_shape[2];
-  int filter_width = filter_shape[3];
+  size_t filter_height = filter_shape[2];
+  size_t filter_width = filter_shape[3];
   // output
   const Shape& output_shape = output->shape();
-  int output_channel = output_shape[1];
-  int output_height = output_shape[2];
-  int output_width = output_shape[3];
+  size_t output_channel = output_shape[1];
+  size_t output_height = output_shape[2];
+  size_t output_width = output_shape[3];
 
-  int batch_input_offset = 0;
-  int batch_output_offset = 0;
-  int dim_left = output_channel;
-  int dim_right = input_channel * filter_height * filter_width;
-  int dim_common = output_height * output_width;
+  size_t batch_input_offset = 0;
+  size_t batch_output_offset = 0;
+  size_t dim_left = output_channel;
+  size_t dim_right = input_channel * filter_height * filter_width;
+  size_t dim_common = output_height * output_width;
 #ifdef BLITZ_DEVELOP
   LOG(INFO) << "dim left: " << dim_left;
   LOG(INFO) << "dim right: " << dim_right;
@@ -273,7 +273,7 @@ void Backend<GPUTensor, DType>::Convolution2DUpdateFunc(
       unpack->data(), const_cast<DType*>(input->data()));
     // Transpose to IC * IH * IW * batch_size
   } else {
-    for (int batch_index = 0; batch_index < batch_size; ++batch_index) {
+    for (size_t batch_index = 0; batch_index < batch_size; ++batch_index) {
       #ifdef BLITZ_PERFORMANCE
       cudaEventRecord(start);
       #endif
